Fixed some incorrect casts to int where it should have been double or float.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@15458 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/store/dbms/rdbms/oracle/Clib/oracle.c
@@ -834,9 +834,9 @@ int ora_get_integer_data (int no_desc, int i)
 	return (int) def[i-1].int_buf;
 }
 
-float ora_get_float_data (int no_desc, int i)
+double ora_get_float_data (int no_desc, int i)
 {
-	return (float) def[i-1].flt_buf;
+	return (double) def[i-1].flt_buf;
 }
 
 float ora_get_real_data (int no_desc, int i)
